Enable warnings as error on Linux (#85)
This commit increases warning level and enables warnings as error on
Linux build. Doing this pointed out several cases of sign mismatches,
variable defintion ordering issues, and unused variables.

+++ cli/gcdmodel.cpp
@@ -114,7 +114,7 @@ bool CGcdData::FixParamOrder( IN Model* submodel )
             {
                 auto p = _modelData.FindParameterByGcdPointer( param );
                 assert( p != _modelData.Parameters.end() );
-                if( p->Order != UNDEFINED_ORDER )
+                if( p->Order != static_cast<unsigned int>(UNDEFINED_ORDER) )
                 {
                     // TODO: add verification of Order
                     // if p->Order > model->parameters.count - model.ResultParameters.count then error out
